Changed and increased BitReader's buffer, minor fixes

+++ TIK2/Encoder.cs
@@ -151,7 +151,7 @@ namespace TIK2
                 var readBuffer = new byte[chunkSize];
                 var len = fs.Length;
 
-                for (int i = 0; i < len; i += chunkSize)
+                for (long i = 0; i < len; i += chunkSize)
                 {
                     if (token.IsCancellationRequested)
                     {
